init-dice-root-ca: Move attestation verification to generated script.
Generating an archive with the attestation is nice but this leaves the
verification up to the user. Since this script had already implemented
a minimal verification process this commit moves this logic into a
generated script. The generated verification script is then run by
init-dice-root-ca.sh to do the verification before returning a
successful exit code. The verification script is included in the archive
of artifacts and recepients of the archive can use it to evaluate the
attestation for themselves (part of building trust in the CA).