Add note in example script identifying need to persist certs.
The dice-ca-init.sh script can only persist the self signed cert for the
root CA. Intermediate CAs are signed by some other CA that may or may
not be on the same platform, yubikey etc. When the intermediate cert is
signed it must be returned to the intermediate, persisted in the
yubikey.

We also need to persist this cert on disk so that the
`send-intermediate` subcommand of `dice-mfg` can send it to devices
being manufactured. We could have this subcommand pull the cert from the
yubikey in the future if necessary but there's likely non-zero overhead
and files on disk is sufficient for current needs.