replace intermediate CA with a mock persistent identity CA
The intermediate in this script was mostly an exercise in building a CA
structure that resembled the reality I was expecting / predicting. Now
that we need another link in the chain to mock the persistent identity
cert we can simplify things by just taking over this intermediate for
the purpose.

The self-signed variant similarly has been shifted from the deviceid to
the intermediate.