dice-mfg: Get passwd from env, allow caller to set auth-id used.
The object-id for the auth credential (auth-id) and the associated
password get combined as a hack to plumb this data through the PKCS#11
auth APIs. It would have been more simple for us to just have whatever
sets the password in the environment to include the object-id of the
auth token too. Somehow passing the auth-id and then combining them
programatically felt more natural.