Add new platform id string format: PDV2.
This format replaces all characters from the revision field with the 'R'
character. Turns out the revision field is not included in the unique
identifier for a platform. It's possible to simply remove the field but
the CN from our cert is a fixed width so this would leave 4 trailing
'\0' characters. This looks funny when rendered by `openssl` so we just
use `RRR` instead. PDV1 strings will remain unchanged to preserve
existing identity certs.