Add sha3-256 hash to Ack message as integrity check.
This allows us to check that the RoT received any message that responds
with Ack. This includes all of the messages that we care about
including: the platforms unique id, it's identity cert and the cert for
the issuing intermediate CA. All other messages that require an Ack
(Break & Ping) will reply with 32 bytes of 0's.