manufacture: `intermediate-cert` param is now required
Previously if this param was omitted the `manufacture` command would
attempt to find the intermediate CA cert in the place the `oks` software
stores it by convention. This was the right thing to do for FCS but now
that we're attempting to support multiple CA interfaces this behavior
requires knowledge of not just openssl but the specific configuration
and filesystem structure created by OKS.

This commit makes this parameter required removing the OKS / openssl
specific knowledge from the `manufacture` command / interface. This will
break consumers that relied on the old default behavior. This param will
be required until we implement a mechanism to get this info from the CAs
we support.