verifier-cli: Persist artifacts as json instead of raw binary.
This falls short of removing all binary encodings but it makes it
possible for us to use the `verify` command to verify an attestation and
persist the artifacts as JSON. Later we can use the `verify-attestation`
and `verify-cert-chain` commands to re-verify the artifacts persisted by
the `verify` command.

Other commands like `attest` and `log` still need to be updated to use
the JSON encoding.