verifier-cli: Add command to check measurements against corpus.
This requires that the caller provide several artifacts:

The `cert-chain` produced by the `Attest` task. This is required because
it includes the FWID recorded by hubris pre-main. This command doesn't
use this cert chain for any other purpose.

The attestation `log` produced by the `Attest` task. This includes the
digests recorded by the `Attest` task once it's up and running.

A `corpus` of measurements that represent our source of truth.
Measurements from this set are considered 'trusted'. The format of this
input is a prototype and will be replaced with the CoRIM stuff currently
under development.

Measurements are extracted from the `cert-chain` and `log` and combined
into a set. If this set is *not* a subset of the `corpus` we produce an
error. This is a simple and rough aproximation of the appraisal process
and will likely change in the future.