Add explicit build step to build the workspace and no_std libs
The resolver unifies features when building packages in the workspace.
This prevents a single invocation of `cargo build --workspace` from
building 'no_std' versions of the libraries here. Adding an explicit
build command for our libraries that have 'std' features ensures that we
catch errors that break the build if 'std' is not enabled.