verifier-cli: strip the interface down to the minimum required
The *_len, cert, and record commands were used to exercise the API but
they're not required in the commandline interface. These are still used
behind the scenes to do allocations. The README relied on using the
`cert` command to get the leaf cert for the attestation signing key but
we can get that from the cert chain w/ a bit of awk.