verifier: Move PkiPath verification into a function.
This was previously exposed through the `PkiPathSignatureVerifier` type
but we'll need to add addtional checks. This new function gives us space
to implement this w/o forcing consumers to juggle a bunch of different
types.

Additionally we update the PkiPathSignatureVerifier type to hold a
reference to the root cert so we don't have to clone it or take
ownership of it.