Move hiffy & ipcc Attest implementations to `dice-verifier`
It would have been nice to preserve the pseudo-factory method in the
`impl dyn Attest` block but moving this to `dice-verifier` would have
made the `Interface` enum part of the API and this enum is part of our
clap interface. It can't remain in `verifier-cli` though since the trait
will now come from the `dice-verifier` library. A simple function can
accomplish the same goal though. If consumers of `dice-verifier` want
this feature we can add it later.