Update `verify_cert_chain` function to verify against multiple roots.
We expect that in the common case consumers will have more than one
trust anchor that they trust. The previous interface required that
consumers verify the cert chain against each root independently. This
works but it's inefficient as it requires redundant verification of the
signatures over the whole chain, not just the roots. This change makes
our API more efficient by accepting multiple roots as an optional slice
of `Certificate` instances. Doing so required redesigning the recursion
strategy: the base, recursion and termination cases.