Add support for v2 SNs
The initial version of the `PlatformId` type and the various conversion
from the 0xv1 / 0xv2 forms made a lot of simplifying (aka bad)
assumptions. Now that we're moving to a second version of the serial
number field these assumptions are getting in the way. This is a quick
fix to get the v2 SNs supported though this code could use a lot of
cleanup.