barcode: Implement a barcode parsing library as a new crate
This is a re-write of the implementation from the
`dice-mfg-msgs::PlatformId` type. It has been refactored into a separate
`Barcode` type. This type is constructed from an `&str` and holds slices
of / references to the various components of the barcode. This allows us
to validate barcodes in a `no_std` environment.

`Barcode` can be used to construct other types whose lifetimes are not
tied to the input string. The `BaseboardId` added in this commit is such
a type. It owns strings that hold both the part and serial numbers and
we use the `Barcode` type to to extract these fields in its constructor.